Fixing location of create user story button.
