+++ sbs-project-management/ui/src/app/epic/epic-base/epic-base.component.css
@@ -1,4 +1,8 @@
-
+.new-user-story {
+  margin-right: 20px;
+  width: 142px;
+  margin-left: auto;
+}
 
 .sidenav {
     height: 100%;
