+++ sbs-project-management/ui/src/app/epic/epic-base/epic-base.component.html
@@ -1,6 +1,3 @@
 <div class="sidenav">
   <a routerLink="/epics/list">Epic List</a>
 </div>
-<div class="main">
-  <router-outlet></router-outlet>
-</div>
