+++ sbs-project-management/ui/src/app/product-backlog/product-backlog.component.html
@@ -11,6 +11,9 @@
 
 <div class="titlebar">
   <h1>Product Backlog</h1>
+  <button class="new-user-story" routerLink="/user-story">
+    New User Story
+  </button>
 </div>
 
 <!-- List to display user stories from productBacklogData -->
