+++ sbs-project-management/ui/src/app/product-backlog/product-backlog.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-product-backlog',
@@ -6,6 +7,7 @@ import { Component } from '@angular/core';
   styleUrls: ['./product-backlog.component.css'],
 })
 export class ProductBacklogComponent {
+  constructor(private router: Router) {}
   // Replace list of user story data obtained from Taiga API in productBacklogData.
   productBacklogData = [
     {
@@ -21,4 +23,7 @@ export class ProductBacklogComponent {
       status: 'Ready to Test',
     },
   ];
+  navigateToUserStoryPage() {
+    this.router.navigate(['/user-story']);
+  }
 }
