+++ sbs-project-management/ui/src/app/project-dashboard/project/project.component.html
@@ -1,5 +1,6 @@
 <div class="sidenav">
   <a [routerLink]="['epics']">Epic List</a>
+  <a [routerLink]="['backlog']">Backlog</a>
 </div>
 <div class="main">
   <div class="accordion">
