+++ sbs-project-management/ui/src/app/project-dashboard/titlebar/titlebar.component.html
@@ -1,8 +1,5 @@
 <div class="titlebar">
   <h1>My Projects</h1>
-
   <button routerLink="/projects/new">New Project</button>
-
-  <button routerLink="/user-story" id="user-story-button">User Story</button>
 </div>
 <app-card></app-card>
